+++ mlu_op/op_src_code/rgb2rgb_resize/env.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-export NEUWARE_HOME=/usr/local/neuware
-export PATH=/usr/local/neuware/bin:$PATH
-export LD_LIBRARY_PATH=/usr/local/neuware/lib64:$LD_LIBRARY_PATH